requirements.txt
bcrypt
beautifulsoup4
certifi
cffi
chardet
cryptography
Django
django-common-helpers
django-cron
django-crontab
django-dbbackup
django-debug-toolbar
django-modelcluster
django-modeltranslation
django-overextends
django-redis
django-rosetta
django-storages
django-taggit
django-treebeard
djangorestframework
draftjs-exporter
dropbox
html5lib
idna
l18n
paramiko
Pillow
ply
polib
psycopg2
pycparser
PyNaCl
python-dotenv
pytz
redis
requests
six
soupsieve
sqlparse
stone
Unidecode
urllib3
uWSGI
wagtail
wagtail-modeltranslation
webencodings
Willow
XlsxWriter
